requirements.txt
torchvision>=0.2.2
pretrainedmodels==0.7.4
torchnet==0.0.4
efficientnet-pytorch==0.4.0
